added redis for caching latest exchange rates

+++ src/main/java/com/example/gateway/JsonApi/model/ClientRequestDTO.java
@@ -17,5 +17,5 @@ public class ClientRequestDTO {
     String timestamp;
     String client;
     String currency;
-    int period;
+    long period;
 }
